Refactor: 필요없는 클래스들 삭제
기존 코드 중 필요없는 것들 모두 삭제함

Related to: #4
