+++ src/main/java/com/hf/healthfriend/auth/config/SecurityConfig.java
@@ -1,0 +1,71 @@
+package com.hf.healthfriend.auth.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.server.resource.introspection.OpaqueTokenIntrospector;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
+
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+public class SecurityConfig {
+    private static final String[] WHITE_LIST = {
+            "/error",
+            "/oauth/code/*",
+            "/swagger-ui/**",
+
+            // TODO: 해당 endpoint 확인 후 삭제할 수 있음
+            "/login",
+            "/api/member/**",
+            "/api/jwt/reissue",
+            "/v3/**"
+    };
+
+    private final OpaqueTokenIntrospector opaqueTokenIntrospector;
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        return
+                http
+                        .cors(corsCustomizer ->corsCustomizer.configurationSource(corsConfigurationSource()))
+                        .csrf(AbstractHttpConfigurer::disable) // TODO: 추후 보안 정책에 따라 CSRF 방지 활용 가능
+                        .formLogin(AbstractHttpConfigurer::disable)
+                        .httpBasic(AbstractHttpConfigurer::disable)
+                        .authorizeHttpRequests((auth) -> auth
+                                .requestMatchers(WHITE_LIST).permitAll()
+                                .anyRequest().authenticated()
+                        )
+                        .oauth2ResourceServer((oauth) ->
+                                oauth.opaqueToken((opaqueToken) ->
+                                        opaqueToken.introspector(this.opaqueTokenIntrospector)))
+                        .sessionManagement((session) -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                        .build();
+    }
+
+    @Bean
+    CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowedOrigins(Arrays.asList("http://localhost:3000"));
+        // config.addAllowedOriginPattern("http://localhost:3000/reissue");
+        config.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        config.setAllowCredentials(true);
+        config.setAllowedHeaders(Arrays.asList("Authorization", "Content-Type", "Accept"));
+        config.setExposedHeaders(Arrays.asList("Authorization"));
+        config.setMaxAge(3600L); // 1시간
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", config);
+        return source;
+    }
+}
+
