--- src/main/java/com/hf/healthfriend/auth/itself/dto/request/LoginRequestDto.java
@@ -1,11 +1,0 @@
-package com.hf.healthfriend.auth.itself.dto.request;
-
-import lombok.Getter;
-
-@Deprecated
-@Getter
-public class LoginRequestDto {
-    private String email;
-    private String password;
-
-}
