--- src/main/java/com/hf/healthfriend/auth/itself/filter/CustomJsonUserPasswordAuthenticationFilter.java
@@ -1,53 +1,0 @@
-package com.hf.healthfriend.auth.itself.filter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hf.healthfriend.auth.itself.dto.request.LoginRequestDto;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.AuthenticationServiceException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.util.StreamUtils;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-@Deprecated
-public class CustomJsonUserPasswordAuthenticationFilter extends AbstractAuthenticationProcessingFilter {
-
-    private final ObjectMapper objectMapper;
-
-    private static final String DEFAULT_LOGIN_REQUEST_URL = "/api/member/login";
-    private static final String HTTP_METHOD = "POST";
-    private static final String CONTENT_TYPE = "application/json";
-    private static final AntPathRequestMatcher DEFAULT_LOGIN_REQUEST_MATCHER
-            = new AntPathRequestMatcher(DEFAULT_LOGIN_REQUEST_URL, HTTP_METHOD);
-
-    public CustomJsonUserPasswordAuthenticationFilter(ObjectMapper objectMapper) {
-        super(DEFAULT_LOGIN_REQUEST_MATCHER);
-        this.objectMapper = objectMapper;
-    }
-
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException, IOException, ServletException {
-        if (request.getContentType() == null || !request.getContentType().equals(CONTENT_TYPE)) {
-            throw new AuthenticationServiceException("Request Content-Type이 application/json 여야 합니다  : " + request.getContentType());
-        }
-
-        String jsonMessageBody = StreamUtils.copyToString(request.getInputStream(), StandardCharsets.UTF_8);
-        LoginRequestDto loginRequest = objectMapper.readValue(jsonMessageBody, LoginRequestDto.class);
-
-        String email = loginRequest.getEmail();
-        String password = loginRequest.getPassword();
-
-        UsernamePasswordAuthenticationToken authenticationToken =
-                new UsernamePasswordAuthenticationToken(email, password, AuthorityUtils.createAuthorityList("ROLE_USER", "ROLE_TEST"));
-
-        return this.getAuthenticationManager().authenticate(authenticationToken);
-    }
-}
