--- src/main/java/com/hf/healthfriend/auth/itself/handler/LoginFailureHandler.java
@@ -1,29 +1,0 @@
-package com.hf.healthfriend.auth.itself.handler;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-/**
- * TODO: 이 코드는 나중에 분석 후 @Deprecated 딱지를 떼거나 아님 걍 참고만 하거나 할 예정
- */
-@Deprecated
-@Component
-@Slf4j
-public class LoginFailureHandler extends SimpleUrlAuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
-        response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-        response.setCharacterEncoding("UTF-8");
-        response.setContentType("text/plain;charset=UTF-8");
-        response.getWriter().write("로그인 실패! 이메일이나 비밀번호를 확인 해 주세요");
-        log.info("로그인에 실패하셨습니다. 메시지 : {}", exception.getMessage());
-    }
-}
