--- src/main/java/com/hf/healthfriend/auth/itself/service/LoginService.java
@@ -1,30 +1,0 @@
-package com.hf.healthfriend.auth.itself.service;
-
-import com.hf.healthfriend.auth.principal.PrincipalDetails;
-import com.hf.healthfriend.domain.member.entity.Member;
-import com.hf.healthfriend.domain.member.repository.MemberRepository;
-import com.hf.healthfriend.global.exception.custom.MemberException;
-import com.hf.healthfriend.global.exception.errorcode.CommonErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Deprecated
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class LoginService implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberException(CommonErrorCode.NO_EXIST_EMAIL_MEMBER_EXCEPTION));
-
-        return new PrincipalDetails(member);
-    }
-}
