+++ src/main/java/com/hf/healthfriend/auth/oauth2/constant/AuthServer.java
@@ -1,0 +1,6 @@
+package com.hf.healthfriend.auth.oauth2.constant;
+
+public enum AuthServer {
+    KAKAO,
+    GOOGLE
+}
