--- src/main/java/com/hf/healthfriend/auth/oauth2/constants/AuthServer.java
@@ -1,6 +1,0 @@
-package com.hf.healthfriend.auth.oauth2.constants;
-
-public enum AuthServer {
-    KAKAO,
-    GOOGLE
-}
