+++ src/main/java/com/hf/healthfriend/auth/oauth2/dto/response/GrantedTokenInfo.java
@@ -1,6 +1,6 @@
 package com.hf.healthfriend.auth.oauth2.dto.response;
 
-import com.hf.healthfriend.auth.oauth2.constants.AuthServer;
+import com.hf.healthfriend.auth.oauth2.constant.AuthServer;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
