--- src/main/java/com/hf/healthfriend/auth/oauth2/dto/response/Oauth2UserInfo.java
@@ -1,12 +1,0 @@
-package com.hf.healthfriend.auth.oauth2.dto.response;
-
-@Deprecated
-public interface Oauth2UserInfo {
-    String getProvider();
-
-    String getProviderId();
-
-    String getEmail();
-
-    String getName();
-}
