+++ src/main/java/com/hf/healthfriend/auth/oauth2/introspect/delegator/KakaoOpaqueTokenIntrospectorDelegator.java
@@ -1,9 +1,9 @@
 package com.hf.healthfriend.auth.oauth2.introspect.delegator;
 
-import com.hf.healthfriend.auth.oauth2.constants.AuthServer;
+import com.hf.healthfriend.auth.oauth2.constant.AuthServer;
 import com.hf.healthfriend.auth.oauth2.dto.response.TokenValidationInfo;
+import com.hf.healthfriend.auth.oauth2.principal.SingleAuthorityOAuth2Principal;
 import com.hf.healthfriend.auth.oauth2.tokenprovider.OAuth2TokenSupport;
-import com.hf.healthfriend.auth.principal.SingleAuthorityOAuth2Principal;
 import com.hf.healthfriend.domain.member.entity.Member;
 import com.hf.healthfriend.domain.member.repository.MemberRepository;
 import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
