+++ src/main/java/com/hf/healthfriend/domain/member/controller/MemberController.java
@@ -1,17 +1,8 @@
 package com.hf.healthfriend.domain.member.controller;
 
-import com.hf.healthfriend.auth.itself.dto.request.SignUpRequestDto;
 import com.hf.healthfriend.domain.member.service.MemberService;
-import com.hf.healthfriend.global.spec.ApiBasicResponse;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -20,34 +11,4 @@ import org.springframework.web.bind.annotation.RestController;
 public class MemberController {
 
     private final MemberService memberService;
-
-    @GetMapping("/reissue")
-    public ResponseEntity<ApiBasicResponse> reissue(
-            @CookieValue(value = "Authorization-refresh") String refreshToken){
-        ResponseEntity<ApiBasicResponse> response = memberService.reissue(refreshToken);
-
-        return response;
-    }
-
-    @GetMapping("/check-duplicate")
-    public ResponseEntity<ApiBasicResponse> checkDuplicate(
-            @RequestParam String email
-    ){
-        ApiBasicResponse response = memberService.checkDuplicateEmail(email);
-
-        return ResponseEntity.status(response.getCode()).body(response);
-    }
-
-    @PostMapping("/sign-up")
-    public ResponseEntity<ApiBasicResponse> signUpMember(
-            @RequestBody @Valid SignUpRequestDto signUpRequestDto
-    ){
-        ApiBasicResponse response = memberService.signUp(signUpRequestDto);
-
-        return ResponseEntity.status(response.getCode()).body(response);
-    }
-
-
-
-
 }
