+++ src/main/java/com/hf/healthfriend/domain/member/repository/MemberRepository.java
@@ -8,7 +8,6 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<Member, String> {
 
     Optional<Member> findByEmail(String email);
-    boolean existsByEmail(String email);
-    Optional<Member> findByRefreshToken(String refreshToken);
 
+    boolean existsByEmail(String email);
 }
