--- src/main/java/com/hf/healthfriend/domain/profile/dto/request/NicknameDto.java
@@ -1,8 +1,0 @@
-package com.hf.healthfriend.domain.profile.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class NicknameDto {
-    private String newNickname;
-}
