--- src/main/java/com/hf/healthfriend/domain/profile/entity/FitnessLevel.java
@@ -1,10 +1,0 @@
-package com.hf.healthfriend.domain.profile.entity;
-
-@Deprecated
-public enum FitnessLevel {
-    /**
-     * 운동 고수, 새싹
-     */
-    BEGINNER,
-    ADVANCED
-}
