--- src/main/java/com/hf/healthfriend/domain/profile/entity/FitnessStyle.java
@@ -1,9 +1,0 @@
-package com.hf.healthfriend.domain.profile.entity;
-
-@Deprecated
-public enum FitnessStyle {
-    /**
-     * 차후 추가 및 영문명으로 변경
-     */
-    소규모형, 그룹형, 의욕만렙형
-}
